[refactor]: Viewpage의 각 영역 grpahql query Converter 클래스로 분리
