+++ src/main/java/synopsis/graphql/service/ScsService.java
@@ -18,7 +18,6 @@ import synopsis.graphql.model.dto.request.CustomScsRequestBody;
 import synopsis.graphql.model.dto.request.RequestScsData;
 import synopsis.graphql.model.dto.request.RequestScsPpvProduct;
 import synopsis.graphql.model.scs.ScsResult;
-import synopsis.graphql.util.converter.JsonToObjectConverter;
 import synopsis.graphql.util.converter.ScsJsonToObjectConverter;
 
 import java.util.Collections;
@@ -36,7 +35,7 @@ public class ScsService {
 
     private final ScsConfig scsConfig;
     private final RestTemplate restTemplate;
-    private final JsonToObjectConverter<ScsResult> scsConverter = new ScsJsonToObjectConverter();
+    private final ScsJsonToObjectConverter scsConverter;
 
 
     private CustomScsPpvProducts toCustomPpvProducts(RequestScsPpvProduct product) {
