+++ src/main/java/synopsis/graphql/service/SmdService.java
@@ -15,7 +15,6 @@ import synopsis.graphql.excpetion.ResultDataNotFoundException;
 import synopsis.graphql.excpetion.SmdRequestException;
 import synopsis.graphql.model.dto.request.RequestSmdData;
 import synopsis.graphql.model.smd.SmdResult;
-import synopsis.graphql.util.converter.JsonToObjectConverter;
 import synopsis.graphql.util.converter.SmdJsonToObjectConverter;
 
 @Slf4j
@@ -25,7 +24,7 @@ public class SmdService {
 
     private final SmdConfig smdConfig;
     private final RestTemplate restTemplate;
-    private final JsonToObjectConverter<SmdResult> smdConverter = new SmdJsonToObjectConverter();
+    private final SmdJsonToObjectConverter smdConverter;
 
 
     public SmdResult getSmdResult(RequestSmdData requestSmdData) {
