+++ src/main/java/synopsis/graphql/util/converter/viewpage/ViewpageConverter.java
@@ -1,0 +1,5 @@
+package synopsis.graphql.util.converter.viewpage;
+
+public interface ViewpageConverter<T> {
+    T convert(Object... sources);
+}
